requirements.txt
python-telegram-bot[webhooks]
python-dotenv
pydantic
mongoengine
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
